storage: support staged addition of layers
The extracting of the tar under the store lock is a bottleneck as many
concurrent processes might hold the locks for a long time on big layers.

To address this move the layer extraction before we take the locks if
possible. Currently this only work when using the overlay driver as the
implementation requires driver specifc details in order for a rename()
to work.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>
(cherry picked from commit c4432727b1b977f3b4f13ede5fe2bce2e68aae6d)
Signed-off-by: Paul Holzinger <pholzing@redhat.com>