In compareImageDestinationManifestEqual, compare full manifests
... instead of digests, making this comparison completely
resistant to cryptographic hash function collisions.

This also means we no longer ignore v2s1 signatures in the comparison;
that is unlikely to matter (given decreasing support for v2s1),
but only skipping the copy if the destination _exactly_ matches
the source, including that signature, is more correct.

(Note that we still return copySingleImageResult.manifestDigest;
this PR does not change that part.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>