Add internal/digests.Options and internal/digests.Options.Choose
The goal is to allow building the digest-choice-dependent machinery
from the bottom up without committing to an API while we don't understand
the problem space; for now, we don't expose any way for users to
actually make a choice.

This code is never called at this point, so this should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>